Add linters and update readme
