--- MyList.rb
@@ -1,13 +1,0 @@
-require('./MyEnumerable')
-
-class MyList
-    include MyEnumerable
-
-    def initialize(*list)
-        @list = list
-    end
-
-    def each               
-        @list.each { |item| yield(item) if block_given? }
-    end
-end
