+++ my_list.rb
@@ -1,0 +1,13 @@
+require('./MyEnumerable')
+
+class MyList
+  include MyEnumerable
+
+  def initialize(*list)
+    @list = list
+  end
+
+  def each
+    @list.each { |item| yield(item) if block_given? }
+  end
+end
